+++ client/app/pods/components/auto-suggest/component.js
@@ -112,6 +112,18 @@ export default Ember.Component.extend({
     });
   },
 
+  /**
+   *  Unique documenet keyup event name for component instance.
+   *  Don't use this before the component is in the DOM
+   *
+   *  @method getKeyupEventName
+   *  @return {String}
+   *  @public
+  **/
+  getKeyupEventName() {
+    return 'keyup.autosuggest-' + this.$().id;
+  },
+
   _resultTextChanged: Ember.observer('resultText', function() {
     if(this.get('searchAllowed')) {
       Ember.run.debounce(this, this.search, this.get('debounce'));
@@ -121,7 +133,7 @@ export default Ember.Component.extend({
   }),
 
   _setupKeybindings: Ember.on('didInsertElement', function() {
-    $(document).on('keyup.autosuggest', (event) => {
+    $(document).on(this.getKeyupEventName(), (event) => {
       if (event.which === 27) {
         this.set('highlightedItem', null);
       }
@@ -138,6 +150,10 @@ export default Ember.Component.extend({
         this.set('searchResults', null);
       }
     });
+  }),
+
+  _teardownKeybindings: Ember.on('willDestroyElement', function() {
+    $(document).off(this.getKeyupEventName());
   }),
 
   selectItem(item) {